pyproject.toml
[project]
name = "tts-extension"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.5",
    "openai-whisper>=20250625",
    "pyperclip>=1.11.0",
    "pynput>=1.8.1",
    "PyYAML>=6.0.3",
    "sounddevice>=0.5.3",
    "typer>=0.20.0",
    "torch>=2.9.1",
]

[project.optional-dependencies]
dev = ["pytest>=7.4", "types-PyYAML>=6.0.12.12"]

[project.scripts]
tts-extension = "tts_extension.cli:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
